+++ api/api/src/main/kotlin/nl/tno/federated/api/Server.kt
@@ -2,6 +2,8 @@ package nl.tno.federated.api
 
 import nl.tno.federated.api.event.type.EventTypeMappingConfig
 import nl.tno.federated.api.graphdb.GraphDBSPARQLClient
+import nl.tno.federated.api.user.UserMappingConfig
+import nl.tno.federated.api.webhook.WebHookKeyStoreConfig
 import org.springframework.boot.SpringApplication
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.context.properties.EnableConfigurationProperties
@@ -17,7 +19,7 @@ import org.springframework.scheduling.annotation.EnableScheduling
 /**
  * FEDeRATED Node API Spring Boot application.
  */
-@EnableConfigurationProperties(EventTypeMappingConfig::class)
+@EnableConfigurationProperties(EventTypeMappingConfig::class, UserMappingConfig::class, WebHookKeyStoreConfig::class)
 @EnableJdbcRepositories
 @EnableScheduling
 @SpringBootApplication(scanBasePackages = ["nl.tno.federated.api"])
