+++ api/api/src/main/kotlin/nl/tno/federated/api/user/UserMappingConfig.kt
@@ -1,8 +1,6 @@
 package nl.tno.federated.api.user
 
 import org.springframework.boot.context.properties.ConfigurationProperties
-import org.springframework.core.io.DefaultResourceLoader
-import java.nio.charset.StandardCharsets
 
 @ConfigurationProperties(prefix = "federated.node.user.admin")
 class UserMappingConfig(val adminUser: AdminUser) {
